+++ src/main/java/com/MooBoo/MooBoo_Spring/adapter/dto/Result.java
@@ -1,4 +1,4 @@
-package com.MooBoo.MooBoo_Spring;
+package com.MooBoo.MooBoo_Spring.adapter.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
