+++ src/main/java/com/MooBoo/MooBoo_Spring/adapter/port/outbound/external/dto/BookApiResponse.java
@@ -1,4 +1,4 @@
-package com.MooBoo.MooBoo_Spring.bookapi.domain;
+package com.MooBoo.MooBoo_Spring.adapter.port.outbound.external.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
@@ -19,7 +19,5 @@ public class BookApiResponse {
     public String query;
     @JsonProperty("item")
     public List<BookItem> item;
-
-
 
 }
