+++ src/main/java/com/MooBoo/MooBoo_Spring/application/port/inbound/BookApiService.java
@@ -1,7 +1,7 @@
-package com.MooBoo.MooBoo_Spring.bookapi.controller.port;
+package com.MooBoo.MooBoo_Spring.application.port.inbound;
 
-import com.MooBoo.MooBoo_Spring.bookapi.domain.BookApi;
-import com.MooBoo.MooBoo_Spring.bookapi.domain.SearchParam;
+import com.MooBoo.MooBoo_Spring.domain.bookapi.BookApi;
+import com.MooBoo.MooBoo_Spring.adapter.port.inbound.api.dto.SearchParam;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
