+++ src/main/java/com/MooBoo/MooBoo_Spring/application/port/outbound/external/BookApiClient.java
@@ -1,7 +1,7 @@
-package com.MooBoo.MooBoo_Spring.bookapi.service.port;
+package com.MooBoo.MooBoo_Spring.application.port.outbound.external;
 
-import com.MooBoo.MooBoo_Spring.bookapi.domain.BookApi;
-import com.MooBoo.MooBoo_Spring.bookapi.domain.SearchParam;
+import com.MooBoo.MooBoo_Spring.domain.bookapi.BookApi;
+import com.MooBoo.MooBoo_Spring.adapter.port.inbound.api.dto.SearchParam;
 import reactor.core.publisher.Mono;
 
 import java.util.List;
