+++ src/main/java/com/MooBoo/MooBoo_Spring/common/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.MooBoo.MooBoo_Spring.config;
+package com.MooBoo.MooBoo_Spring.common.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
